+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.2.4] - 2025-09-26
 ### Added
-- added rules: `acceptedIf()`
+- updated: Rules can accept closures
+- added rules: `acceptedIf()`,`size()`,`extensions()`
 - updated: `distinct()`,`url()`,`email()`,`timezone()`,`alphabetAndNumeric()`
 
 ## [1.2.3] - 2025-09-24